# integrations/csm-integration-service/openapi.yaml
openapi: 3.0.1
info:
  title: CSM Integration Service API
  version: 1.0.0
  description: >
    Project/account search and their Contacts sub-resource, incident
    creation, and alert-incident mapping create/lookup, for third-party
    (M2M) consumers. Fronted by Choreo's API Manager gateway (subscription +
    client credentials) — this service performs no inbound auth of its own.
security:
  - oauth2ClientCredentials: []
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /accounts/{id}:
    get:
      summary: Get an account by ID.
      operationId: getAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The account with the given ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        "400":
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized (upstream reported an authorization failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/search:
    post:
      summary: Search accounts.
      operationId: searchAccounts
      requestBody:
        description: Account search payload
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountSearchPayload'
      responses:
        "200":
          description: Accounts matching the search query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSearchResponse'
        "400":
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized (upstream reported an authorization failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/{id}/contacts/search:
    post:
      summary: Search contacts for an account.
      operationId: searchAccountContacts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Account contact search payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountContactSearchPayload'
      responses:
        "200":
          description: Contacts for the given account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountContactSearchResponse'
        "400":
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized (upstream reported an authorization failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    oauth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          # Placeholder — fill in with the actual token endpoint for this service's
          # Choreo-managed API once it's provisioned. Not a real/guessed URL.
          tokenUrl: "<token_url>"
          scopes: {}
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          description: Number of results per page.
        offset:
          type: integer
          description: Zero-based offset into the result set.
    Account:
      type: object
      properties:
        id:
          type: string
        sfId:
          type: string
        name:
          type: string
        tier:
          type: string
          enum: [basic, enterprise]
        region:
          type: string
        activationDate:
          type: string
          format: date-time
        deactivationDate:
          type: string
          format: date-time
        ownerId:
          type: string
        technicalOwnerId:
          type: string
        agentEnabled:
          type: boolean
        kbReferencesEnabled:
          type: boolean
        createdOn:
          type: string
          format: date-time
        updatedOn:
          type: string
          format: date-time
    AccountSearchPayload:
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        filters:
          type: object
          properties:
            searchQuery:
              type: string
              description: Case-insensitive match against account name or Salesforce ID.
    AccountSearchResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    AccountContactSearchPayload:
      type: object
      properties:
        filters:
          type: object
          properties:
            searchQuery:
              type: string
              description: Searches across contact name and email (case-insensitive).
        pagination:
          $ref: '#/components/schemas/Pagination'
    AccountContactSearchResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              email:
                type: string
              isPrimary:
                type: boolean
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer
...